# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Brand & Marketing OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ColorPalette {
  primary: string;
  secondary: string;
  accent: string;
  background: string;
  text: string;
}

export interface BrandKit {
  personality: string;
  positioning: string;
  toneOfVoice: string;
  audienceSegments: string[];
  visualStyle: string;
  colorPalette: ColorPalette;
  visualStyleRules: string;
}

export interface BrandSummary {
  id: number;
  companyName: string;
  industry: string;
  /** @nullable */
  logoUrl?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface Brand {
  id: number;
  companyName: string;
  companyDescription: string;
  industry: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  status: string;
  brandKit?: BrandKit | null;
  createdAt: string;
  updatedAt: string;
}

export interface GenerateBrandKitBody {
  /** Extracted hex colors from the uploaded logo */
  brandColors?: string[];
}

export interface GenerateCampaignBody {
  /** User instructions for the campaign (tone, goals, style, target) */
  brief?: string;
  /** Base64 data URLs of reference images to inspire visuals */
  referenceImages?: string[];
  /**
   * Number of posts to generate (defaults to 7)
   * @minimum 1
   * @maximum 14
   */
  postCount?: number;
}

export interface CreateBrandBody {
  companyName: string;
  companyDescription: string;
  industry: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** Extracted hex colors from the uploaded logo */
  brandColors?: string[];
}

export interface UpdateBrandBody {
  companyName?: string;
  companyDescription?: string;
  industry?: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
}

export interface CampaignDay {
  day: number;
  objective: string;
  postConcept: string;
  marketingAngle: string;
  cta: string;
}

export interface SocialPost {
  id: number;
  campaignId: number;
  day: number;
  caption: string;
  hook: string;
  cta: string;
  hashtags: string[];
  imagePrompt: string;
  /**
   * Base64 data URL of the AI-generated image
   * @nullable
   */
  imageUrl?: string | null;
  platform: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdatePostBody {
  caption?: string;
  hook?: string;
  cta?: string;
  hashtags?: string[];
  imagePrompt?: string;
  platform?: string;
}

export interface Campaign {
  id: number;
  brandId: number;
  title: string;
  strategy: string;
  days: CampaignDay[];
  posts: SocialPost[];
  createdAt: string;
  updatedAt: string;
}

export interface BrandStats {
  brandId: number;
  totalCampaigns: number;
  totalPosts: number;
  brandKitGenerated: boolean;
  /** @nullable */
  lastCampaignDate?: string | null;
}

export interface DashboardSummary {
  totalBrands: number;
  totalCampaigns: number;
  totalPosts: number;
  recentBrands: BrandSummary[];
}